+++ support/network/network.go
@@ -12,12 +12,15 @@ package network
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
+	"net"
 	"os"
+	"strings"
 
 	"github.com/essentialkaos/ek/v12/netutil"
 	"github.com/essentialkaos/ek/v12/req"
 	"github.com/essentialkaos/ek/v12/sliceutil"
 	"github.com/essentialkaos/ek/v12/sortutil"
+	"github.com/essentialkaos/ek/v12/strutil"
 
 	"github.com/essentialkaos/ek/v12/support"
 )
@@ -65,9 +68,9 @@ func cleanIPList(ips []string) []string {
 }
 
 // resolvePublicIP resolves public IP using IP resolver
-func resolvePublicIP(resolver string) string {
+func resolvePublicIP(resolverURL string) string {
 	resp, err := req.Request{
-		URL:         resolver,
+		URL:         resolverURL,
 		AutoDiscard: true,
 	}.Get()
 
@@ -75,5 +78,44 @@ func resolvePublicIP(resolver string) string {
 		return ""
 	}
 
-	return resp.String()
+	var ip string
+
+	if strings.HasSuffix(resolverURL, "/cdn-cgi/trace") {
+		ip = extractIPFromCloudflareTrace(resp.String())
+	} else {
+		ip = resp.String()
+	}
+
+	if isValidIP(ip) {
+		return ip
+	}
+
+	return ""
+}
+
+// extractIPFromCloudflareTrace extracts public IP from Cloudflare trace
+// response
+func extractIPFromCloudflareTrace(data string) string {
+	for i := 0; i < 16; i++ {
+		f := strutil.ReadField(data, i, false, '\n')
+
+		if f == "" {
+			break
+		}
+
+		n := strutil.ReadField(f, 0, false, '=')
+
+		if n != "ip" {
+			continue
+		}
+
+		return strutil.ReadField(f, 1, false, '=')
+	}
+
+	return ""
+}
+
+// isValidIP returns if given IP is valid
+func isValidIP(ip string) bool {
+	return ip != "" && net.ParseIP(ip) != nil
 }